Move actions to a new component
