+++ js/components/HomeScene/Scene/Scene.jsx
@@ -1,5 +1,4 @@
 import { useEffect, useRef, useState } from "react";
-import * as THREE from "three";
 import { useThree } from "@react-three/fiber";
 import {
     CameraControls,
@@ -11,6 +10,7 @@ import { Perf } from "r3f-perf";
 
 import { debug } from "@/js/core/constants";
 import { SceneContext } from "./SceneContext";
+import Actions from "./Actions";
 import Model from "./Model";
 
 const padding = 0.5;
@@ -22,98 +22,24 @@ export default function Scene() {
     const camera = useRef(null);
     const cameraControls = useRef(null);
     const triptychRef = useRef(null);
-    const currentSubjectRef = useRef(null);
-
-    /**
-     *
-     * @param {'right'|'up'|'down'|'left'} direction
-     */
-    async function orbit(direction = "right") {
-        const cameraPosition = new THREE.Vector3();
-        cameraPosition.copy(camera.current.position);
-        console.log(cameraPosition);
-
-        currentSubjectRef.current.geometry.computeBoundingSphere();
-        const boundingSphere =
-            currentSubjectRef.current.geometry.boundingSphere.clone();
-
-        const boundingBox = new THREE.Box3();
-        boundingSphere.getBoundingBox(boundingBox);
-
-        const currentSubjectPosition = new THREE.Vector3();
-        currentSubjectPosition.copy(currentSubjectRef.current.position);
-
-        switch (direction) {
-            case "up": {
-                await cameraControls.current.rotate(
-                    0,
-                    THREE.MathUtils.degToRad(-90),
-                    true,
-                );
-                await cameraControls.current.fitToBox(
-                    currentSubjectRef.current,
-                    true,
-                    {
-                        paddingTop: padding,
-                        paddingRight: padding,
-                        paddingBottom: padding,
-                        paddingLeft: padding,
-                    },
-                );
-                break;
-            }
-            case "down": {
-                await cameraControls.current.rotate(
-                    0,
-                    THREE.MathUtils.degToRad(90),
-                    true,
-                );
-                await cameraControls.current.fitToBox(
-                    currentSubjectRef.current,
-                    true,
-                    {
-                        paddingTop: padding,
-                        paddingRight: padding,
-                        paddingBottom: padding,
-                        paddingLeft: padding,
-                    },
-                );
-                break;
-            }
-            case "left": {
-                await cameraControls.current.rotate(
-                    THREE.MathUtils.degToRad(-45),
-                    THREE.MathUtils.degToRad(cameraPosition.y > 0 ? 45 : -45),
-                    true,
-                );
-                break;
-            }
-            case "right":
-            default: {
-                await cameraControls.current.rotate(
-                    THREE.MathUtils.degToRad(45),
-                    THREE.MathUtils.degToRad(cameraPosition.y > 0 ? 45 : -45),
-                    true,
-                );
-                break;
-            }
-        }
-    }
-
+    const currentSubject = useRef(null);
 
     useEffect(() => {
         setMounted(true);
-        currentSubjectRef.current = triptychRef.current;
+        currentSubject.current = triptychRef.current;
         cameraControls.current.setOrbitPoint(
-            currentSubjectRef.current.x,
-            currentSubjectRef.current.y,
-            currentSubjectRef.current.z,
+            currentSubject.current.x,
+            currentSubject.current.y,
+            currentSubject.current.z,
         );
     }, []);
 
     useEffect(() => {
         if (!mounted) return;
-        cameraControls.current.fitToBox(currentSubjectRef.current, true, {
+        if (!debug) {
+            cameraControls.current.disconnect();
+        }
+        cameraControls.current.fitToBox(currentSubject.current, true, {
             paddingTop: padding,
             paddingRight: padding,
             paddingBottom: padding,
@@ -122,11 +48,12 @@ export default function Scene() {
     }, [mounted, width, height, cameraControls]);
 
     return (
-        <SceneContext.Provider value={{ cameraControls, camera }}>
+        <SceneContext.Provider
+            value={{ cameraControls, camera, currentSubject, padding }}
+        >
             {/* debug */}
             {debug && <Perf position="top-left" />}
             {debug && <axesHelper args={[5]} />}
-            {debug && <OrbitControls />}
 
             {/* camera */}
             <OrthographicCamera
@@ -137,16 +64,12 @@ export default function Scene() {
             <CameraControls ref={cameraControls} />
 
             {/* environment */}
-            <Environment preset="sunset" background blur={1} />
-            <directionalLight
-                castShadow
-                position={[-1.5, 2.5, 1.8]}
-                intensity={1.5}
-            />
-            <ambientLight intensity={0.5} />
+            <Environment preset="sunset" blur={1} />
 
             {/* models */}
             <Model ref={triptychRef} />
+
+            <Actions />
         </SceneContext.Provider>
     );
 }
